+++ extract_msg/attachments/custom_att_handler/__init__.py
@@ -21,6 +21,7 @@ CustomAttachmentHandler and add it using the `registerHandler` function.
 __all__ = [
     # Classes.
     'CustomAttachmentHandler',
+    'JournalAssociatedAttachment',
     'OutlookImageDIB',
 
     # Functions.
@@ -56,6 +57,7 @@ def registerHandler(handler : Type[CustomAttachmentHandler]) -> None:
 # Import built-in handler modules. They will all automatically register their
 # respecive handler(s).
 from .outlook_image_dib import OutlookImageDIB
+from .jrnl_assoc_att import JournalAssociatedAttachment
 
 
 if TYPE_CHECKING: