+++ CHANGELOG.md
@@ -11,6 +11,11 @@
 * Fixed issues with `MessageBase.date` and related things either being incorrectly documented or doing things that are not specified by the documentation. It was *supposed* to have been changed to use `datetime` objects, but it was still using strings.
 * Removed unused function `extract_msg.utils.isEmptyString`.
 * Removed unused function `extract_msg.utils.properHex`.
+* Added a `getStream` helper function to `CustomAttachmentHandler`.
+* Added a `getStreamAs` helper function to `CustomAttachmentHandler`.
+* Added new custom attachment handler for journal-associated attachments.
+* Changed `EntryID.autoCreate` to return `None` if given `None` or empty bytes.
+* Changed `EntryID.autoCreate` to raise a `FeatureNotImplemented` exception if no valid entry ID class is found.
 
 **v0.45.0**
 * BREAKING: Changed parsing of string multiple properties to remove the trailing null byte. This *will* cause the output of parsing them to differ.